requirements/base.txt
mysqlclient==1.3.12
anyjson==0.3.3
cycler==0.10.0
Django==1.10
matplotlib==2.1.2
numpy==1.14.0
pbkdf2==1.3
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
six==1.11.0
xlrd==0.9.0
django-uuslug==1.1.8
django-filebrowser-no-grappelli==3.7.0
Pillow==4.1.1
psycopg2==2.7.4
django-modeladmin-reorder==0.2
singlemodeladmin==0.7

